requirements.txt
numpy==1.18.1
pandas==1.0.1
matplotlib == 3.1.3
scikit-learn==0.23.1
joblib == 0.14.1
tpot == 0.11.1
pillow==7.0.0 
streamlit == 0.61.0